pyproject.toml
[project]
name = "km-mind-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "fastmcp>=2.13.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
